requirements.txt
streamlit
requests
selenium
webdriver-manager
pandas
python-dotenv